requirements.txt
Django==1.7.8
Fabric
Jinja2
MarkupSafe
Pillow
Pygments
Sphinx
Whoosh
amqp
anyjson
asana
billiard
celery
django-appconf
django-autoslug
django-celery
django-compressor
django-cors-headers
django-guardian
django-gunicorn
django-haystack
django-kombu
django-libsass
django-mail-queue
django-mptt
django-polymorphic
djangorestframework
django-celery-email
django-filter
docutils
drf-nested-routers
ecdsa
factory-boy
gunicorn
kombu
libsass
paramiko
pycrypto
pyparsing
python-magic
python-ptrace
pytz
requests
six
sorl-thumbnail
